最新合并

+++ main.cpp
@@ -1,7 +1,12 @@
 #include "common.h"
 
-bool debug_mode = false;
+bool debug_mode = true;
 bool debug_mode_mark_disk_imfromation = false;
+
+bool jump_1_round = false;
+std::string history_name = ".\\history\\1.txt"; // 跳过 1 时加载的历史文件名，如果没有跳过，这也是保存名
+bool save_round_1 = false;
+
 
 int main()
 {
@@ -12,32 +17,58 @@ int main()
     }
     
     auto start_time = std::chrono::high_resolution_clock::now();
-    pre_process();
-    global_turn = 1; // 第一轮
-    for (t = 1, time_step = 1; t <= T_time_step_length + EXTRA_TIME; t++,time_step++) {
-        
-        if(t % FRE_PER_SLICING == 1){
-            if(debug_mode){
-                freopen("CON", "w", stdout);
-                printf("current time step %d   %d   %d\n", t, zero_request, selected_r);
-                std::cout << "  empty_object_read: " << empty_object_read << std::endl;
-                std::cout << "  empty_request_read: " << empty_request_read << std::endl;
-                std::cout << "  effective_read: " << effective_read << std::endl;
-                std::cout << "  all_finish_select_size: " << all_finish_select_size << std::endl;
-                std::cout << "  all_finish_request_efficiency: " << std::fixed << all_finish_request_efficiency << std::endl;
-                std::cout << "  select_zero_request: " << select_zero_request << std::endl;
-                std::cout << "  all_finish_select: " << all_finish_select << std::endl;
-                std::cout << "  select_but_not_finish: " << select_but_not_finish << std::endl;
-                std::cout << "  drop_req_num: " << drop_req_num << std::endl;
-                std::cout << "  all_mark: " << all_mark << std::endl;
-                freopen(".\\output.txt", "a+", stdout);
+
+    if(!jump_1_round || !debug_mode){ // 跳过第一轮或者非 debug 模式
+        pre_process();
+        global_turn = 1; // 第一轮
+        for (t = 1, time_step = 1; t <= T_time_step_length + EXTRA_TIME; t++,time_step++) {
+            
+            if(t % FRE_PER_SLICING == 1){
+                if(debug_mode){
+                    freopen("CON", "w", stdout);
+                    printf("current time step %d   %d   %d\n", t, zero_request, selected_r);
+                    std::cout << "  empty_object_read: " << empty_object_read << std::endl;
+                    std::cout << "  empty_request_read: " << empty_request_read << std::endl;
+                    std::cout << "  effective_read: " << effective_read << std::endl;
+                    std::cout << "  all_finish_select_size: " << all_finish_select_size << std::endl;
+                    std::cout << "  all_finish_request_efficiency: " << std::fixed << all_finish_request_efficiency << std::endl;
+                    std::cout << "  select_zero_request: " << select_zero_request << std::endl;
+                    std::cout << "  all_finish_select: " << all_finish_select << std::endl;
+                    std::cout << "  select_but_not_finish: " << select_but_not_finish << std::endl;
+                    std::cout << "  drop_req_num: " << drop_req_num << std::endl;
+                    std::cout << "  all_mark: " << all_mark << std::endl;
+                    freopen(".\\output.txt", "a+", stdout);
+                }
+            }
+            
+            time_step_action();
+            
+            delete_action();
+           
+            write_action();
+            
+            read_action();
+            
+            exchange_action();
+        }
+        int n;
+        scanf("%d",&n);
+        for(int i=0;i<n;i++){
+            int id,tag;
+            scanf("%d%d", &id, &tag);
+            object_array[id].tag = tag;
+            object_array[id].true_tag = true;
+            for(int j=0;j<obj_read_data[id].size();j++){
+                tag_read[tag][j] += obj_read_data[id][j];
             }
         }
-        time_step_action();
-        delete_action();
-        write_action();
-        read_action();
-        exchange_action();
+        if(!save_round_1 || !debug_mode){
+
+        }else{ // 保存第一轮 且 debug 模式
+            save_history();
+        }
+    }else{
+        load_history();
     }
 
     // 第二轮
@@ -46,10 +77,13 @@ int main()
     select_but_not_finish=0;
     drop_req_num=0;
     max_segment_select_size=4;
+    double efficient_disk_rate = 0.38;
     segment_num=14;
     int efficient_size = ceil((double)V_block_per_disk * efficient_disk_rate);
     segment_size = ceil((double)efficient_size / (double)segment_num);
     efficient_disk_end = segment_size * segment_num;
+    selected_r = 0;
+    un_selected_r = 0;
     pre_process_2();
 
     // 根据上一轮的全局信息进行统计
@@ -73,10 +107,15 @@ int main()
                 freopen(".\\output.txt", "a+", stdout);
             }
         }
+        
         time_step_action();
+        
         delete_action();
+       
         write_action();
+        
         read_action();
+        
         exchange_action();
     }
